23RD Aug Updated Code

+++ app/src/main/java/online/internship/CartAdapter.java
@@ -25,7 +25,7 @@ public class CartAdapter extends RecyclerView.Adapter<CartAdapter.MyHolder> {
     public CartAdapter(Context context, ArrayList<CartList> arrayList) {
         this.context = context;
         this.arrayList = arrayList;
-        sp = context.getSharedPreferences(ConstantSp.PREF,Context.MODE_PRIVATE);
+        sp = context.getSharedPreferences(ConstantSp.PREF, Context.MODE_PRIVATE);
 
         db = context.openOrCreateDatabase("Online_Internship", Context.MODE_PRIVATE, null);
 
@@ -49,8 +49,8 @@ public class CartAdapter extends RecyclerView.Adapter<CartAdapter.MyHolder> {
 
     public class MyHolder extends RecyclerView.ViewHolder {
 
-        ImageView imageView,deleteIv;
-        TextView name,price,total;
+        ImageView imageView, deleteIv, add, remove;
+        TextView name, price, total, qty;
 
         public MyHolder(@NonNull View itemView) {
             super(itemView);
@@ -59,6 +59,11 @@ public class CartAdapter extends RecyclerView.Adapter<CartAdapter.MyHolder> {
             name = itemView.findViewById(R.id.custom_cart_name);
             price = itemView.findViewById(R.id.custom_cart_price);
             total = itemView.findViewById(R.id.custom_cart_total);
+
+            add = itemView.findViewById(R.id.custom_cart_plus);
+            remove = itemView.findViewById(R.id.custom_cart_minus);
+            qty = itemView.findViewById(R.id.custom_cart_qty);
+
         }
     }
 
@@ -66,16 +71,19 @@ public class CartAdapter extends RecyclerView.Adapter<CartAdapter.MyHolder> {
     public void onBindViewHolder(@NonNull CartAdapter.MyHolder holder, int position) {
         holder.imageView.setImageResource(Integer.parseInt(arrayList.get(position).getProductImage()));
         holder.name.setText(arrayList.get(position).getProductName());
-        holder.price.setText(ConstantSp.PRICE_SYMBOL+arrayList.get(position).getProductPrice() +" * "+arrayList.get(position).getProductQty());
-        holder.total.setText(ConstantSp.PRICE_SYMBOL+arrayList.get(position).getTotalPrice());
+        holder.price.setText(ConstantSp.PRICE_SYMBOL + arrayList.get(position).getProductPrice());
+        holder.total.setText(ConstantSp.PRICE_SYMBOL + arrayList.get(position).getTotalPrice());
+
+        holder.qty.setText(arrayList.get(position).getProductQty());
+
         holder.itemView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                sp.edit().putString(ConstantSp.PRODUCT_ID,arrayList.get(position).getProductId()).commit();
-                sp.edit().putString(ConstantSp.PRODUCT_NAME,arrayList.get(position).getProductName()).commit();
+                sp.edit().putString(ConstantSp.PRODUCT_ID, arrayList.get(position).getProductId()).commit();
+                sp.edit().putString(ConstantSp.PRODUCT_NAME, arrayList.get(position).getProductName()).commit();
                 sp.edit().putInt(ConstantSp.PRODUCT_IMAGE, Integer.parseInt(arrayList.get(position).getProductImage())).commit();
-                sp.edit().putString(ConstantSp.PRODUCT_PRICE,arrayList.get(position).getProductPrice()).commit();
-                sp.edit().putString(ConstantSp.PRODUCT_DESC,arrayList.get(position).getProductDesc()).commit();
+                sp.edit().putString(ConstantSp.PRODUCT_PRICE, arrayList.get(position).getProductPrice()).commit();
+                sp.edit().putString(ConstantSp.PRODUCT_DESC, arrayList.get(position).getProductDesc()).commit();
                 new CommonMethod(context, ProductDetailActivity.class);
             }
         });
@@ -83,23 +91,104 @@ public class CartAdapter extends RecyclerView.Adapter<CartAdapter.MyHolder> {
         holder.deleteIv.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                String deleteQuery = "DELETE FROM CART WHERE CARTID='"+arrayList.get(position).getCartId()+"'";
+                String deleteQuery = "DELETE FROM CART WHERE CARTID='" + arrayList.get(position).getCartId() + "'";
                 db.execSQL(deleteQuery);
-                new CommonMethod(context,"Product Removed From Cart");
+                new CommonMethod(context, "Product Removed From Cart");
 
                 CartFragment.iTotalPrice -= Integer.parseInt(arrayList.get(position).getTotalPrice());
-                CartFragment.checkout.setText("Checkout "+ConstantSp.PRICE_SYMBOL+CartFragment.iTotalPrice);
+                CartFragment.checkout.setText("Checkout " + ConstantSp.PRICE_SYMBOL + CartFragment.iTotalPrice);
 
-                if(CartFragment.iTotalPrice==0){
+                if (CartFragment.iTotalPrice == 0) {
                     CartFragment.dataLayout.setVisibility(View.GONE);
                     CartFragment.emptyLayout.setVisibility(View.VISIBLE);
-                }
-                else{
+                } else {
                     CartFragment.dataLayout.setVisibility(View.VISIBLE);
                     CartFragment.emptyLayout.setVisibility(View.GONE);
                 }
 
                 arrayList.remove(position);
+                notifyDataSetChanged();
+
+            }
+        });
+
+        holder.add.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                int iQty = Integer.parseInt(arrayList.get(position).getProductQty()) + 1;
+                int iTotalPrice = Integer.parseInt(arrayList.get(position).getProductPrice()) * iQty;
+
+                holder.qty.setText(String.valueOf(iQty));
+                holder.price.setText(ConstantSp.PRICE_SYMBOL + arrayList.get(position).getProductPrice());
+                holder.total.setText(ConstantSp.PRICE_SYMBOL + iTotalPrice);
+
+                CartList list = new CartList();
+                list.setCartId(arrayList.get(position).getCartId());
+                list.setProductId(arrayList.get(position).getProductId());
+                list.setProductName(arrayList.get(position).getProductName());
+                list.setProductImage(arrayList.get(position).getProductImage());
+                list.setProductDesc(arrayList.get(position).getProductDesc());
+                list.setProductPrice(arrayList.get(position).getProductPrice());
+                list.setProductQty(String.valueOf(iQty));
+                list.setTotalPrice(String.valueOf(iTotalPrice));
+                arrayList.set(position, list);
+
+                CartFragment.iTotalPrice += Integer.parseInt(arrayList.get(position).getProductPrice());
+                CartFragment.checkout.setText("Checkout " + ConstantSp.PRICE_SYMBOL + CartFragment.iTotalPrice);
+
+                CartFragment.dataLayout.setVisibility(View.VISIBLE);
+                CartFragment.emptyLayout.setVisibility(View.GONE);
+
+                String updateQuery = "UPDATE CART SET PRODUCTQTY='"+iQty+"',TOTALPRICE='"+iTotalPrice+"' WHERE CARTID='"+arrayList.get(position).getCartId()+"'";
+                db.execSQL(updateQuery);
+
+                notifyDataSetChanged();
+            }
+        });
+
+        holder.remove.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                int iQty = Integer.parseInt(arrayList.get(position).getProductQty()) - 1;
+                int iTotalPrice = Integer.parseInt(arrayList.get(position).getProductPrice()) * iQty;
+
+                holder.qty.setText(String.valueOf(iQty));
+                holder.price.setText(ConstantSp.PRICE_SYMBOL + arrayList.get(position).getProductPrice());
+                holder.total.setText(ConstantSp.PRICE_SYMBOL + iTotalPrice);
+
+                CartFragment.iTotalPrice -= Integer.parseInt(arrayList.get(position).getProductPrice());
+                CartFragment.checkout.setText("Checkout " + ConstantSp.PRICE_SYMBOL + CartFragment.iTotalPrice);
+
+                if (CartFragment.iTotalPrice == 0) {
+                    CartFragment.dataLayout.setVisibility(View.GONE);
+                    CartFragment.emptyLayout.setVisibility(View.VISIBLE);
+                } else {
+                    CartFragment.dataLayout.setVisibility(View.VISIBLE);
+                    CartFragment.emptyLayout.setVisibility(View.GONE);
+                }
+
+                if(iQty>0) {
+                    CartList list = new CartList();
+                    list.setCartId(arrayList.get(position).getCartId());
+                    list.setProductId(arrayList.get(position).getProductId());
+                    list.setProductName(arrayList.get(position).getProductName());
+                    list.setProductImage(arrayList.get(position).getProductImage());
+                    list.setProductDesc(arrayList.get(position).getProductDesc());
+                    list.setProductPrice(arrayList.get(position).getProductPrice());
+                    list.setProductQty(String.valueOf(iQty));
+                    list.setTotalPrice(String.valueOf(iTotalPrice));
+                    arrayList.set(position, list);
+
+                    String updateQuery = "UPDATE CART SET PRODUCTQTY='"+iQty+"',TOTALPRICE='"+iTotalPrice+"' WHERE CARTID='"+arrayList.get(position).getCartId()+"'";
+                    db.execSQL(updateQuery);
+
+                }
+                else{
+                    String updateQuery = "DELETE FROM CART WHERE CARTID='"+arrayList.get(position).getCartId()+"'";
+                    db.execSQL(updateQuery);
+                    arrayList.remove(position);
+                }
+
                 notifyDataSetChanged();
 
             }
